refactor(Topology): :recycle: Add new tests for topology process details
